+++ Erfpacht058-Web/src/app/app/rapportage/sjablonen/sjablonen.component.ts
@@ -40,7 +40,7 @@ export class SjablonenComponent implements OnInit {
     const dialogRef = this.dialog.open(DeleteDialogComponent);
 
     dialogRef.afterClosed().subscribe(result => {
-      if (result.delete) {
+      if (result && result.delete) {
         this.http.delete('/template/' + id).subscribe(resp => {
           this.getTemplates();
         });
